change: split up the tests into separate files
